+++ packages/react-ui/internal/ThemePlayground/CheckboxPlayground.tsx
@@ -5,7 +5,8 @@ import { Checkbox, CheckboxProps } from '../../components/Checkbox';
 
 import { getComponentsFromPropsList } from './helpers';
 
-const propsList: Array<CheckboxProps & { focused?: boolean }> = [
+type CheckboxProp = CheckboxProps & { focused?: boolean };
+const propsList: CheckboxProp[] = [
   { children: 'Default' },
   { children: 'Checked', checked: true },
   { children: 'Disabled', checked: true, disabled: true },