+++ packages/react-ui/internal/ThemePlayground/ThemeContextPlayground.tsx
@@ -13,6 +13,7 @@ import { ComboBox } from '../../components/ComboBox';
 import { Link } from '../../components/Link';
 import * as ColorFunctions from '../../lib/styles/ColorFunctions';
 import { Writeable } from '../../typings/utility-types';
+import { findPropertyDescriptor } from '../../lib/theming/ThemeHelpers';
 
 import { ThemeEditor } from './ThemeEditor';
 import { styles } from './Playground.styles';
@@ -213,13 +214,4 @@ export class ThemeContextPlayground extends React.Component<PlaygroundProps, Pla
 
     return ThemeFactory.create<ThemeIn>(result);
   };
-}
-
-function findPropertyDescriptor(theme: Theme, propName: keyof Theme) {
-  for (; theme != null; theme = Object.getPrototypeOf(theme)) {
-    if (Object.prototype.hasOwnProperty.call(theme, propName)) {
-      return Object.getOwnPropertyDescriptor(theme, propName) || {};
-    }
-  }
-  return {};
 }