+++ packages/react-ui/lib/__tests__/reactGetTextContent-test.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react/display-name */
 import React from 'react';
 
 import { reactGetTextContent } from '../reactGetTextContent';
@@ -12,10 +13,12 @@ class SimpleComponent extends React.Component {
   }
 }
 
-const testCase: Array<{
+type TextContent = {
   label: string;
   renderNode: (label: string) => React.ReactNode;
-}> = [
+};
+
+const testCase: TextContent[] = [
   {
     label: 'First',
     renderNode: (label) => label,