+++ packages/react-ui/lib/dom/getDOMRect.ts
@@ -1,24 +1,36 @@
+import { Nullable } from '../../typings/utility-types';
+type DOMRectDefaultValues = Omit<DOMRect, 'toJSON'>;
 /**
  * Возвращает размер элемента и его позицию относительно viewport
  *
- * @param ref - ref элемента
+ * @param element - ref элемента или сам элемент
  * @returns - возвращает размер элемента и его позицию относительно viewport
  */
-export const getDOMRect = <T extends HTMLElement>(ref: React.RefObject<T>): Omit<DOMRect, 'toJSON'> => {
-  if (!ref.current) {
-    const defaultValues = {
-      width: 0,
-      height: 0,
-      x: 0,
-      y: 0,
-      bottom: 0,
-      left: 0,
-      right: 0,
-      top: 0,
-    };
-
-    return defaultValues;
+export const getDOMRect = <T extends Element>(element: Nullable<T> | React.RefObject<T>): DOMRectDefaultValues => {
+  const defaultValues: DOMRectDefaultValues = {
+    width: 0,
+    height: 0,
+    x: 0,
+    y: 0,
+    bottom: 0,
+    left: 0,
+    right: 0,
+    top: 0,
+  };
+  if (element && 'current' in element) {
+    return getRefRect(defaultValues, element);
   }
-
-  return ref.current.getBoundingClientRect();
+  return getElementRect(defaultValues, element);
+};
+const getElementRect = (defaultValues: DOMRectDefaultValues, element: Nullable<Element>) => {
+  if (element) {
+    return element.getBoundingClientRect();
+  }
+  return defaultValues;
+};
+const getRefRect = <T extends Element>(defaultValues: DOMRectDefaultValues, ref: React.RefObject<T>) => {
+  if (ref?.current) {
+    return ref.current.getBoundingClientRect();
+  }
+  return defaultValues;
 };