+++ packages/react-ui/lib/events/keyboard/extractCode.ts
@@ -1,7 +1,7 @@
 import { Entries } from '../../../typings/utility-types';
 
 import { KeyboardEventCodes as Codes } from './KeyboardEventCodes';
-import { KeyboardMapKeys, KeyCode, Location } from './KeyboardMapKeys';
+import { KeyboardKey, KeyboardMapKeys } from './KeyboardMapKeys';
 
 export function extractCode(e: React.KeyboardEvent<HTMLElement> | KeyboardEvent): Codes {
   e = (e as React.KeyboardEvent<HTMLElement>).nativeEvent || e;
@@ -10,7 +10,7 @@ export function extractCode(e: React.KeyboardEvent<HTMLElement> | KeyboardEvent)
   }
   const keyCode = e.keyCode || e.which;
   const location = e.location;
-  const mapped = (Object.entries as Entries<Codes, Array<[KeyCode, Location]>>)(KeyboardMapKeys).find(([, set]) =>
+  const mapped = (Object.entries as Entries<Codes, KeyboardKey[]>)(KeyboardMapKeys).find(([, set]) =>
     set.some(([_keyCode, _location]) => _keyCode === keyCode && _location === location),
   );
   return mapped ? mapped[0] : Codes.Unidentified;